fixed - personal access token migration and removed old assets from resources

+++ database/migrations/2023_04_11_192047_add_expires_at_to_personal_access_tokens_table.php
@@ -13,7 +13,9 @@ return new class extends Migration {
     public function up()
     {
         Schema::table('personal_access_tokens', function (Blueprint $table) {
-            $table->timestamp('expires_at')->nullable()->after('last_used_at');
+            if (! Schema::hasColumn('personal_access_tokens', 'expires_at')) {
+                $table->timestamp('expires_at')->nullable()->after('last_used_at');
+            }
         });
     }
 
@@ -25,7 +27,9 @@ return new class extends Migration {
     public function down()
     {
         Schema::table('personal_access_tokens', function (Blueprint $table) {
-            $table->dropColumn('expires_at');
+            if (Schema::hasColumn('personal_access_tokens', 'expires_at')) {
+                $table->dropColumn('expires_at');
+            }
         });
     }
 };
